批改/pyproject.toml
[project]
name = "ai-grading-agent"
version = "0.1.0"
description = "生产级纯视觉 AI 批改系统"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "temporalio>=1.5.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=3.0.0",
    "langchain-google-genai>=1.0.0",
    "fastapi>=0.110.0",
    "pydantic>=2.6.0",
    "psycopg[pool]>=3.1.0",
    "redis>=5.0.0",
    "pdf2image>=1.17.0",
    "imagehash>=4.3.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.9",
    "alembic>=1.13.0",
    "pillow>=10.2.0",
    "pymupdf>=1.26.7",
    "google-generativeai>=0.8.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.98.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
